added separate flight dialog controller
