+++ webapp/controller/Detail.controller.js
@@ -24,30 +24,6 @@ sap.ui.define([
 			// this.oRouter.getRoute("detailDetail").attachMatched(this._onProductMatched, this);
 		},
 
-		onActionAdd: function () {
-			var oView = this.getView();
-
-			// create dialog lazily
-			if (!this.pDialog) {
-				this.pDialog = Fragment.load({
-					id: oView.getId(),
-					name: "tdapp.view.fragment.AddAction",
-					controller : this
-				}).then(function (oDialog) {
-					// connect dialog to the root view of this component (models, lifecycle)
-					oView.addDependent(oDialog);
-					return oDialog;
-				});
-			} 
-			this.pDialog.then(function(oDialog) {
-				oDialog.open();
-			});
-		},
-
-		onDialogCancelPress: function() {
-			this.byId("AddActionDialog").close();
-		},
-
 		/* 
 			get path of selected action
 			pass to detailDetail
