+++ webapp/manifest.json
@@ -8,6 +8,17 @@
 		}
 	},
 	"sap.ui5": {
+		"extends": {
+			"extensions": {
+				"sap.ui.controllerExtenstions":{
+					"tdapp.controller.Detail":{
+						"controllerNames":[
+							"tdapp.controller.extension.AddFlightDialog"
+						]
+					}
+				}
+			}
+		},
 		"rootView": {
 			"viewName": "tdapp.view.App",
 			"type": "XML",
