+++ webapp/source/Action.json
@@ -34,14 +34,5 @@
         "Description": "You can add more text here for additional notes 4",
         "DateSet": "28/06/2021",
         "DateDue": "01/07/2021"
-    },
-    {
-        "ItemId":"a5",
-        "Category": "c2",
-        "Selected": true,
-        "Action": "Another to completed Action",
-        "Description": "You can add more text here for additional notes 5",
-        "DateSet": "28/06/2021",
-        "DateDue": "01/07/2021"
     }
 ]
