+++ webapp/view/fragment/AddAction.fragment.xml
@@ -2,6 +2,7 @@
   xmlns="sap.m"
   xmlns:form="sap.ui.layout.form"
   xmlns:core="sap.ui.core"
+  xmlns:l="sap.ui.layout"
   core:require="{
     'ODataStringType': 'sap/ui/model/odata/type/String',
     'ODataDateTimeType': 'sap/ui/model/odata/type/DateTime',
@@ -31,6 +32,20 @@
         }
       }"
     ></Input>
+    <!-- The commented out below is for selecting which list to add action to
+          currently it is fixed that the action will be added to the To Do List
+          add this functionality in the future -->
+    <!-- <Label text="List"></Label>
+    <MultiComboBox
+			selectionChange="handleSelectionChange"
+			selectionFinish="handleSelectionFinish"
+			width="350px"
+			items="{
+				path: 'categories>/',
+				sorter: { path: 'CategoryName' }
+			}">
+			<core:Item key="{categories>CategoryId}" text="{categories>CategoryName}"/>
+		</MultiComboBox> -->
     <Label text="Additional Notes"></Label>
     <Input
       fieldGroupIds="inputs"
